queue implementation cleanup
